+++ tests/testbed/request.nim
@@ -3,15 +3,15 @@ import ./dataset
 type Request* = ref object
   dataset: Dataset
   id: string
-  duration: int
-  expiry: int
+  duration: uint64
+  expiry: uint64
 
 func init*(
   _: type Request,
   dataset: Dataset,
   id: string,
-  duration: int,
-  expiry: int
+  duration: uint64,
+  expiry: uint64
 ): Request =
   Request(dataset: dataset, id: id, duration: duration, expiry: expiry)
 
@@ -21,8 +21,8 @@ func dataset*(request: Request): Dataset =
 func id*(request: Request): string =
   request.id
 
-func duration*(request: Request): int =
+func duration*(request: Request): uint64 =
   request.duration
 
-func expiry*(request: Request): int =
+func expiry*(request: Request): uint64 =
   request.expiry
